finalizing testing

+++ nasnav-services/src/main/java/com/nasnav/dao/BankAccountActivityRepository.java
@@ -12,7 +12,7 @@ import java.util.List;
 @Repository
 public interface BankAccountActivityRepository extends CrudRepository<BankAccountActivityEntity, Long> {
     @Query("SELECT SUM(m.amountIn) - SUM(m.amountOut) FROM BankAccountActivityEntity m where m.account.id = :accountId and m.id > :startId")
-    Long getBalance(long startId, long accountId);
+    Float getBalance(long startId, long accountId);
     BankAccountActivityEntity findFirstByAccount_IdOrderByIdDesc(Long accountId);
     PageImpl<BankAccountActivityEntity> findAllByAccount_Id(long accountId, Pageable page);
 }
