+++ nasnav-services/src/main/java/com/nasnav/dao/BankReservationRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 @Repository
 public interface BankReservationRepository extends JpaRepository<BankReservationEntity, Long> {
     @Query("SELECT SUM(m.amount) FROM BankReservationEntity m where m.account.id = :accountId and m.fulfilled = false")
-    Long getReservedBalance(long accountId);
+    Float getReservedBalance(long accountId);
     List<BankReservationEntity> getAllByAccount_Id(long accountId);
     List<BankReservationEntity> getAllByAccount_IdAndFulfilled(long accountId, boolean isFulfilled);
 }
