+++ nasnav-services/src/main/java/com/nasnav/dto/response/BankAccountDetailsDTO.java
@@ -6,7 +6,7 @@ import java.time.LocalDateTime;
 
 @Data
 public class BankAccountDetailsDTO extends BankAccountDTO{
-    private Long openingBalance;
+    private Float openingBalance;
     private LocalDateTime openingBalanceDate;
     private Boolean locked;
     private BankBalanceSummaryDTO summary;
