+++ nasnav-services/src/main/java/com/nasnav/service/BankInsideTransactionService.java
@@ -1,5 +1,5 @@
 package com.nasnav.service;
 
 public interface BankInsideTransactionService {
-    public void transfer(long receiverAccountId, long amount);
+    public void transfer(long receiverAccountId, float amount);
 }
