+++ nasnav-services/src/main/java/com/nasnav/service/BankOutsideTransactionService.java
@@ -1,7 +1,7 @@
 package com.nasnav.service;
 
 public interface BankOutsideTransactionService {
-    public void depositOrWithdrawal(long amount, boolean isDeposit, String transactionIdOfBC);
-    public Boolean validateDepositOrWithdrawalIsDone(String hashBC, long amount);
+    public void depositOrWithdrawal(float amount, boolean isDeposit, String transactionIdOfBC);
+    public Boolean validateDepositOrWithdrawalIsDone(String hashBC, float amount);
 
 }
