+++ nasnav-services/src/main/java/com/nasnav/service/BankReservationService.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface BankReservationService {
     //create new reserve
-    public BankReservationDTO createReservation(long accountId, long amount);
+    public BankReservationDTO createReservation(long accountId, float amount);
     public List<BankReservationDTO> getReservations(Boolean isFulfilled);
     public BankReservationDTO getReservationById(long reservationId);
     //fulfil record
